Speed up getBlockInfo
Added a new variable to the feed variable which stores copies of the first and last (theoretically) stopTimes to make the .filter call used in getBlockInfo faster by searching over objects we don't care about.

Vercel shuts down your applet when it's not used so accessing the site constantly fires the GTFS process which is an intense process usually ran at the start of the program and when an update for the file is found (both rare occurrences). Change was made to accommodate vercel.